회원 정보 수정 시 주소 정보도 함께 업데이트 한다

+++ domain/src/main/java/sion/bookstore/domain/member/repository/Address.java
@@ -8,8 +8,8 @@ import java.util.Date;
 @Getter
 @Setter
 public class Address {
-    private Integer id;
-    private Integer memberId;
+    private Long id;
+    private Long memberId;
     private String name;
     private String addressBasic;
     private String addressDetail;
